requirements.txt
streamlit
pandas
python-dotenv
matplotlib
requests
openpyxl
